+++ Unity/Packages/com.greenbox.i18n/Editor/Usage/Analysis/AssemblyInfo.cs
@@ -4,3 +4,4 @@ using System.Runtime.CompilerServices;
 [assembly: InternalsVisibleTo("GreenBox.I18n.Usage.Index")]
 [assembly: InternalsVisibleTo("GreenBox.I18n.Unity.Editor")]
 [assembly: InternalsVisibleTo("GreenBox.I18n.Development.Editor")]
+[assembly: InternalsVisibleTo("GreenBox.I18n.Unity.Editor.Tests")]